requirements.txt
djangorestframework
drf-yasg
mysqlclient
django-filter
python-decouple